Added basic unit tests for each individual state in the application

+++ Glass Size Estimator/Glass Size Estimator/States/TruncateState.cs
@@ -6,24 +6,24 @@ using System.Threading.Tasks;
 
 namespace Glass_Size_Estimator
 {
-	// Implements truncating logic
-	public class TruncateState : State
-	{
-		// Constructor
-		public TruncateState(int stateNumber, int nextState) : base(stateNumber, nextState)
-		{
-			// Do Nothing
-		}
+    // Implements truncating logic
+    public class TruncateState : State
+    {
+        // Constructor
+        public TruncateState(int stateNumber, int nextState) : base(stateNumber, nextState)
+        {
+            // Do Nothing
+        }
 
-		/*
+        /*
 		 * PROCESS DESCRIPTION: Truncate the given value.
 		 * INPUT: { Current Value (FLOAT) }
 		 * PARAMETERS: { None }
 		 * OUTPUT: { New Value (FLOAT) }
 		 */
-		public override void Process()
-		{
-			this.Output = Math.Truncate((float)this.Input);
-		}
-	}
+        public override void Process()
+        {
+            this.Output = (float)Math.Truncate((float)this.Input);
+        }
+    }
 }
